+++ OUDS/Core/Themes/Orange/Sources/Providers/ComponentTokens/OrangeThemeButtonComponentTokensProvider.swift
@@ -11,16 +11,17 @@
 // Software description: A SwiftUI components library with code examples for Orange Unified Design System
 //
 
+import OUDS
 import OUDSFoundations
 import OUDSTokensSemantic
 
-/// A class which wraps all **component  tokens of buttons** for button objects like `OUDSButton`.
+/// A class which wraps all **component  tokens of buttons** for *button* objects like `OUDSButton`.
 /// Contains also references to semantic tokens providers so as to be able to use them to define the component tokens.
 /// This provider should be integrated as a `AllButtonComponentTokensProvider` implementation inside `OUDSTheme` so as to provide
 /// all tokens to the users. It helps users to override some of the tokens and assign them to an `OUDSTheme` implementation to use.
 /// Custom themes can use subclass of ``OrangeThemeButtonComponentTokensProvider`` and apply the provider they need.
-/// It implements also the protocol `ButtonComponentTokens` so as to expose the component tokens for buttons through any `OUDSTheme`.
-/// Button components tokens are defined with raw and semantic tokens of sizes (from `AllSizeSemanticTokensProvider`), borders
+/// It implements also the protocol `ButtonComponentTokens` so as to expose the component tokens for *buttons* through any `OUDSTheme`.
+/// *Button* components tokens are defined with raw and semantic tokens of sizes (from `AllSizeSemanticTokensProvider`), borders
 /// (from `AllBorderSemanticTokensProvider`), colors (from `AllColorSemanticTokensProvider`) and
 /// spaces (from `AllSpaceSemanticTokensProvider`).
 ///
@@ -58,7 +59,7 @@ import OUDSTokensSemantic
 /// class LocalTheme: OrangeTheme {
 ///
 ///     override init() {
-///         super.init(button: CustomButtonComponentTokensProvider(),
+///         super.init(tokensProviders: [ CustomButtonComponentTokensProvider(), ... ])
 ///     }
 /// }
 /// ```
@@ -66,11 +67,11 @@ import OUDSTokensSemantic
 /// or to an already existing theme for example:
 ///
 /// ```swift
-///     OrangeTheme(button: CustomButtonComponentTokensProvider())
+///     OrangeTheme(tokensProviders: [ CustomButtonComponentTokensProvider(), ... ])
 /// ```
 ///
 /// - Since: 0.9.0
-open class OrangeThemeButtonComponentTokensProvider {
+open class OrangeThemeButtonComponentTokensProvider: AllButtonComponentTokensProvider {
 
     /// Provider of size semantic tokens to use for button sizes
     public let sizes: AllSizeSemanticTokensProvider