+++ OUDS/Core/Themes/Orange/Sources/Providers/ComponentTokens/OrangeThemeSelectComponentTokensProvider.swift
@@ -11,16 +11,17 @@
 // Software description: A SwiftUI components library with code examples for Orange Unified Design System
 //
 
+import OUDS
 import OUDSFoundations
 import OUDSTokensSemantic
 
-/// A class which wraps all **component  tokens of select** for select objects like `OUDSSelect`.
+/// A class which wraps all **component  tokens of select** for *select* objects like `OUDSSelect`.
 /// Contains also references to semantic tokens providers so as to be able to use them to define the component tokens.
 /// This provider should be integrated as a `AllSelectComponentTokensProvider` implementation inside `OUDSTheme` so as to provide
 /// all tokens to the users. It helps users to override some of the tokens and assign them to an `OUDSTheme` implementation to use.
 /// Custom themes can use subclass of ``OrangeThemeSelectComponentTokensProvider`` and apply the provider they need.
-/// It implements also the protocol `SelectComponentTokens` so as to expose the component tokens for links through any `OUDSTheme`.
-/// Link components tokens are defined with raw and semantic tokens of sizes (from `AllSizeSemanticTokensProvider`),
+/// It implements also the protocol `SelectComponentTokens` so as to expose the component tokens for *select* through any `OUDSTheme`.
+/// *Select* components tokens are defined with raw and semantic tokens of sizes (from `AllSizeSemanticTokensProvider`),
 /// colors (from `AllColorSemanticTokensProvider`) and spaces (from `AllSpaceSemanticTokensProvider`).
 ///
 /// ```swift
@@ -53,7 +54,7 @@ import OUDSTokensSemantic
 /// class LocalTheme: OrangeTheme {
 ///
 ///     override init() {
-///         super.init(select: CustomSelectComponentTokensProvider())
+///         super.init(tokensProviders: [ CustomSelectComponentTokensProvider(), ... ])
 ///     }
 /// }
 /// ```
@@ -61,11 +62,11 @@ import OUDSTokensSemantic
 /// or to an already existing theme for example:
 ///
 /// ```swift
-///     OrangeTheme(select: CustomSelectComponentTokensProvider())
+///     OrangeTheme(tokensProviders: [ CustomSelectComponentTokensProvider(), ... ])
 /// ```
 ///
 /// - Since: 0.9.0
-open class OrangeThemeSelectComponentTokensProvider {
+open class OrangeThemeSelectComponentTokensProvider: AllSelectComponentTokensProvider {
 
     /// Provider of size semantic tokens to use for link sizes
     public let sizes: AllSizeSemanticTokensProvider