+++ OUDS/Core/Themes/Orange/Sources/Providers/SemanticTokens/OrangeThemeGridSemanticTokensProvider.swift
@@ -11,6 +11,7 @@
 // Software description: A SwiftUI components library with code examples for Orange Unified Design System 
 //
 
+import OUDS
 import OUDSFoundations
 
 /// A class which wraps all **grid semantic tokens** and expose them.
@@ -43,7 +44,7 @@ import OUDSFoundations
 /// class LocalTheme: OrangeTheme {
 ///
 ///     override init() {
-///         super.init(grids: CustomGridTokensProvider(),
+///         super.init(tokensProviders: [ CustomGridTokensProvider(), ... ])
 ///     }
 /// }
 /// ```
@@ -51,11 +52,11 @@ import OUDSFoundations
 /// or to an already existing theme for example:
 ///
 /// ```swift
-///     OrangeTheme(grids: CustomGridTokensProvider())
+///     OrangeTheme(tokensProviders: [ CustomGridTokensProvider(), ... ])
 /// ```
 ///
 /// - Since: 0.8.0
-open class OrangeThemeGridSemanticTokensProvider {
+open class OrangeThemeGridSemanticTokensProvider: AllGridSemanticTokensProvider {
 
     /// Intializes the provider
     public init() {